% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Importance score computation for object JRF}
\usage{
importance(x, scale = TRUE)
}
\arguments{
\item{x}{object of class JRF}

\item{scale}{either TRUE or FALSE}
}
\value{
\code{"imp"} return importance scores (total decrease in node impurity)
}
\description{
Importance score computation. This file is a modified version of function Importance contained in the R package randomForest.
}

\references{
A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. \emph{R News} \bold{2}, 18--22.
}
