% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Importance score computation for object JRF
(This file is a modified version of file importance.R contained in the R package randomForest.)}
\usage{
importance(x, scale = TRUE)
}
\arguments{
\item{x}{object of class JRF}

\item{scale}{either TRUE or FALSE}
}
\value{
imp matrix of importance scores (\code{"imp"})
}
\description{
Importance score computation for object JRF
(This file is a modified version of file importance.R contained in the R package randomForest.)
}

